% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_ize_pffd.R
\name{fst_ize_pffd}
\alias{fst_ize_pffd}
\title{function that fst-izes the output from pairwise fst from dumpfile}
\usage{
fst_ize_pffd(pffd, treelist)
}
\arguments{
\item{pffd}{The output list from \code{\link[=pairwise_fst_from_dumpfile]{pairwise_fst_from_dumpfile()}}.}

\item{treelist}{The named list that tells us who is in which populations.
We need this to know which population each index value in pffd represents.}
}
\description{
blah, blah
}
