% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_ize_zbars.R
\name{fst_ize_zbars}
\alias{fst_ize_zbars}
\title{Internal function to convert the z_bars to pop-specific Fst's}
\usage{
fst_ize_zbars(SI, FET)
}
\arguments{
\item{SI}{the list that gets output by srs_identity.}

\item{FET}{the fst_edges_tibble.  A tibble with columns parent daughter parent_idx and daughter_idx.
each daughter appears once.  The idxs must be the indexes of the columns in the
zbar output of SI that correspond to each parent or daughter node.}
}
\description{
Blah, blah
}
\keyword{internal}
