% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cycles.R
\name{n_cycles}
\alias{n_cycles}
\title{Find directed cycles of a given length in a directed graph}
\usage{
n_cycles(G, n)
}
\arguments{
\item{G}{the directed graph passed in as a list with names as the parents
and daughters as the values.  The graph must be such that each
parent has only a single daughter.  Though a node can be the daughter
of multiple parents.}

\item{n}{the length of the cycles that should be found}
}
\value{
Returns a list with the following elements:
\itemize{
\item \code{length} holds the values of \code{n}
\item \code{cycles} is a list of the cycles in which each cycle is
just the names of the nodes in cycle order, starting with the first one,
lexicographically.
\item \code{remaining} is a list of edges remaining after all of the nodes involved
in the currently-found cycles have been removed.
}
}
\description{
This is intended to be used iteratively: first find all the cycles of
length 2.  Then remove any edges that involve any of the nodes in those
two-cycles.  Then see if there are any three-cycles, etc.
}
