% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_fst_from_dumpfile}
\alias{pairwise_fst_from_dumpfile}
\title{Read the Y dumpfile}
\usage{
pairwise_fst_from_dumpfile(file, num_shorts, L, num_pops)
}
\arguments{
\item{file}{path to the dumpfile}

\item{num_shorts}{the number of unsigned shorts that were written to the dumpfile}

\item{L}{the number of loci expected here}

\item{num_pops}{the number of populations expected.}
}
\value{
This passes back a list of information. more specifics later
}
\description{
Ultimately I will want to compute pairwise pop-specific Fst's
}
