% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_treelist.R
\name{prepare_treelist}
\alias{prepare_treelist}
\title{Prepare a tree of individuals/populations for population-specific Fst calculations}
\usage{
prepare_treelist(treelist, vcf_names = NULL)
}
\arguments{
\item{treelist}{the desired tree, specified as a named list.  Each component
of this list is named for an internal node in the tree, and the contents of that
component is a character vector of the nodes (or leaves/individuals) that are the
daughters of that node. This tree is thought of as a rooted tree, with a topmost node
that is typically called "root", but could be named anything.  The first components
of the list must all the nodes representing populations, whose daughters are
exclusively individuals.  After that, the remaining nodes must be the other
internal nodes in the tree.}

\item{vcf_names}{the names of the samples in the vcf file in the order that their
columns appear in the vcf file.  These don't have to be exactly the names as they
appear in the VCF file.  The important part is that they correspond to the individual
names in treelist and that they are ordered as the individuals appear as columns
in the VCF file. Note that in this current implementation, the columns of the VCF
file must not include any individuals other than those that are in treelist. I might
try to change that later, but have no guarantees.}
}
\description{
This takes a list-based specification of a tree and it: checks it to make sure
it is correct, verifies that it is a tree, makes some plots of it so the
user can ensure they have input the data correctly, identifies the populations
(which are, by definition, one node level above the leaves), and then translates
all the portions of the tree into an integer-based specification that is
appropriate for the \code{\link[=srs_identity]{srs_identity()}} function.
}
