% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib2dot.R
\name{tib2dot}
\alias{tib2dot}
\title{given a tibble specification of the pop-hierarchy, make a dot file of the tree}
\usage{
tib2dot(tib, outfile = "tree.dot", names_as_numbers = FALSE,
  names_on_indivs = FALSE, num_gene_copies = 2, int_fontsize = 50,
  indiv_size = 0.5, edge_extras = NULL)
}
\arguments{
\item{tib}{the tibble}

\item{outfile}{the file to write it out.  By default it is tree.dot}

\item{num_gene_copies}{for each individual at the bottom of the tree,
how many gene copies should come off of him/her}

\item{int_fontsize}{size of font for labels on internal nodes.  User
must fiddle with this to get it right.}

\item{indiv_size}{size of the little circles at the bottom that denote individuals.
User must fiddle with this to get it right.}
}
\description{
given a tibble specification of the pop-hierarchy, make a dot file of the tree
}
