% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srs_identity}
\alias{srs_identity}
\title{Compute relative ibd measures from single read sampling}
\usage{
srs_identity(
  file,
  pops_of_indivs,
  num_pops,
  num_internal_nodes,
  daughters,
  BootReps
)
}
\arguments{
\item{file}{path to the file that holds the Chrom, Pos, \link{AD} file (i.e. allele_depths.txt
in the description above).}

\item{pops_of_indivs}{a vector of 0-based indices of the populations each of the individuals
belongs to. Must be in the order that individuals appear in the VCF file.}

\item{num_pops}{The number of internal nodes in the tree that are directly above the individuals.}

\item{num_internal_nodes}{the number of internal nodes in the tree (this includes the populations)}

\item{daughters}{a list of vectors.  Each one is the 0-based indexes of the daughters
of the non-population internal nodes. Here 0 corresponds to the first internal node (the first population).}
}
\value{
This passes back a list of information about the various allele
sharing statistics (and bootstrap information) that can be
used by a higher-level function to compute the F-statistics, etc.
}
\description{
This implements the calculations from the
"Unified Characterization of Population Structure
and Relatedness" framework of Weir and Goudet (2017), but it does it from
the allele depth information in a VCF file, by sampling a single
read from each individual.
}
\details{
This method only works on biallelic markers (but it can work on biallelic indels
if you have them).  You must do all the filtering you want to do on your VCF file
(using bcftools, for example) and then you have to extract the read depths from that
file and put it into a file on your hard drive somewhere.  That file is the
input to this function.

Note that it would not be too hard to modify this to deal with multiple alleles.

If you had all your data in a VCF file called raw.vcf, here is how you would
process it:
\enumerate{
\item Filter it.  Something like this to get biallelic markers typed at over 50\\%
of individuals (i.e., having at least one read at over 50\\% of individuals), and
with a minor allele frequency of > 0.05:
}

\code{bcftools view -m 2 -M 2 --min-af 0.05 --max-af 0.95 -i 'F_MISSING < 0.5' raw.vcf > filtered.vcf}
\enumerate{
\item Then extract the allele depths out of that. The will look like 0,1 or 2,1 or,
if they are missing, they will just be a dot, ".".   So, each row has 2 + N whitespace
separated values
on it, (where N is the number of individuals). The first two words on each row
are Chrom and Pos.  This assumes that the allele depths
are comma-separated, and that sites missing reads are denoted by a period:
}

\code{bcftools query -f '\%CHROM\t\%POS[\t\%AD]\n' filtered.vcf > allele_depths.txt}
\enumerate{
\item You might want to check that allele_depths has the right number of "words" in it.  So, you
would do \code{wc allele_depths.txt}, and confirm that the second number in the output
is equal to \eqn{(N + 2) * L}, where \eqn{N} is the number of individuals (samples)
in the data set, filtered.vcf, and \eqn{L} is the number of markers in filtered.vcf.
In other words, the file should have L rows, each one with N + 2 columns: 2 for Chrom and
Pos and then 1 column for each of the N individuals.
\item Then, you call this function, giving it the path to the file allele_depths.txt,
and you also pass in a list of vectors that hold information about the
hierarchical levels you want to use to be able to compute relative measures of
identity by descent.
}
}
